`fn get_skip_ctx`: Deduplicate.
The `recon_tmpl_16.rs` version has already had its `MERGE_CTX` macros translated as unsafe Rust macros,
but since I plan to translate them to safe Rust closures, I didn't keep them.
